// src/app/api/generated/fn/tours/search-tours.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { TourSearchResponse } from '../../models/tour-search-response';

export interface SearchTours$Params {
  q?: string;
  transportType?: 'BIKE' | 'HIKE' | 'RUNNING' | 'VACATION';
  popularity?: 'NEW' | 'RARELY_USED' | 'POPULAR' | 'VERY_POPULAR';
  childFriendliness?: 'UNKNOWN' | 'FAMILY_FRIENDLY' | 'MODERATE_FAMILY_SUITABILITY' | 'CHALLENGING_ROUTE' | 'ADULT_ORIENTED';
  ratingMin?: number;
}

export function searchTours(http: HttpClient, rootUrl: string, params?: SearchTours$Params, context?: HttpContext): Observable<StrictHttpResponse<TourSearchResponse>> {
  const rb = new RequestBuilder(rootUrl, searchTours.PATH, 'get');
  if (params) {
    rb.query('q', params.q, {});
    rb.query('transportType', params.transportType, {});
    rb.query('popularity', params.popularity, {});
    rb.query('childFriendliness', params.childFriendliness, {});
    rb.query('ratingMin', params.ratingMin, {});
  }

  return http.request(
    rb.build({ responseType: 'blob', accept: '*/*', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<TourSearchResponse>;
    })
  );
}

searchTours.PATH = '/api/tours';