// src/app/api/generated/models/tour-summary-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ComputedTourAttributesDto } from '../models/computed-tour-attributes-dto';
import { CoverImageDto } from '../models/cover-image-dto';
export interface TourSummaryDto {
  computedAttributes?: ComputedTourAttributesDto;
  coverImage?: CoverImageDto;
  createdAt?: string;
  endLocation?: string;
  estimatedDurationS?: number;
  id?: number;
  name?: string;
  plannedDistanceM?: number;
  startLocation?: string;
  timezoneId?: string;
  transportType?: 'BIKE' | 'HIKE' | 'RUNNING' | 'VACATION';
  updatedAt?: string;
  userId?: number;
}